Markdown: interpolate variables before rendering markdown
...instead of the other way around. Doing markdown first prevents
specifying an alternate variable format in the URL of a link, because
the two ways of doing that (${var:fmt} and [[var:fmt]]) both end up
getting percent-encoded in the process and therefore don't work.

The only other thing I think this could change would be if a variable
expands to some valid markdown, it would cause it to get transformed by
the markdown engine, whereas before it would not. This should be safe,
though, as markdown is a pretty limited set of tags you can expand to.